Építkezés, Támadás véglegesítése

+++ Szakdoga_Unity/Assets/Scripts/ResourceObject.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ResourceObject : MonoBehaviour {
 
-    public int Capacity;
+    public float Capacity;
     public resourceType Type;
 
 }
